requirements.txt
numpy
tensorly
tlviz
plotly
kaleido
ipython
ipykernel
matplotlib
pandas
tqdm
nbformat
shootout-opt
black
-e .

# Also manually install custom tensorly branch
# TODO